% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetups.R
\name{get_joined_meetups}
\alias{get_joined_meetups}
\title{Get Joined Meetups}
\usage{
get_joined_meetups(key, fields = c("id", "name", "urlname", "link"), ...)
}
\arguments{
\item{key}{an API key from https://www.meetup.com/meetup_api/}

\item{fields}{a charcter vector of the fields to return}

\item{...}{a named list where each element is a character vector for additional
parameters e.g. `list("omit" = c("member.photo", "member.event_context")`}
}
\value{
returns a data.frame of the meetups associated to an API key
}
\description{
This request is specific to an API key. Provides urlname used
in other functions.
}
\examples{
\dontrun{
get_joined_meetups("your_api_key")
}

}
