% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance.R
\name{mark_event_attendance}
\alias{mark_event_attendance}
\title{Mark Event Attendance}
\usage{
mark_event_attendance(urlname, event_id, member_ids, member_status, key)
}
\arguments{
\item{urlname}{string, URL name for the meetup. e.g 'R-Users-Sydney'}

\item{event_id}{string, event id for the meetup event.}

\item{member_ids}{vector of member id strings to mark attendance for.}

\item{member_status}{vector string for member status, positions match those in
member_id's.}

\item{key}{an API key from https://www.meetup.com/meetup_api/.}
}
\value{
the POST request object.
}
\description{
Mark Event Attendance
}
\examples{
\dontrun{
mark_event_attendance("R-Users-Sydney", "your_event_id",
                      "my_member_id", "attended",
                      "your_api_key")
}

}
