% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance.R
\name{get_event_attendance}
\alias{get_event_attendance}
\title{Title}
\usage{
get_event_attendance(urlname, event_id, key, fields = c("status", "member",
  "rsvp.response"), ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
